Split automation action slug domains
