+++ src/app_core/native_shell/composition/state/automation/action_slugs/shell_sources.rs
@@ -1,0 +1,76 @@
+use super::super::*;
+
+pub(super) fn action_slug(action: &UiAction) -> Option<&'static str> {
+    Some(match action {
+        UiAction::SelectColumn { .. } => "select_column",
+        UiAction::MoveColumn { .. } => "move_column",
+        UiAction::ToggleTransport => "toggle_transport",
+        UiAction::PlayCompareAnchor => "play_compare_anchor",
+        UiAction::PlayFromStart => "play_from_start",
+        UiAction::PlayFromCurrentPlayhead => "play_from_current_playhead",
+        UiAction::PlayFromWaveformCursor => "play_from_waveform_cursor",
+        UiAction::PlayWaveformAtPrecise { .. } => "play_waveform_at_precise",
+        UiAction::HandleEscape => "handle_escape",
+        UiAction::FocusBrowserPanel => "focus_browser_panel",
+        UiAction::FocusSourcesPanel => "focus_sources_panel",
+        UiAction::FocusWaveformPanel => "focus_waveform_panel",
+        UiAction::FocusFolderPanel => "focus_folder_panel",
+        UiAction::FocusLoadedContentInList => "focus_loaded_content_in_list",
+        UiAction::FocusBrowserSearch => "focus_browser_search",
+        UiAction::BlurBrowserSearch => "blur_browser_search",
+        UiAction::OpenAddSourceDialog => "open_add_source_dialog",
+        UiAction::OpenOptionsMenu => "open_options_menu",
+        UiAction::CloseOptionsPanel => "close_options_panel",
+        UiAction::EditDefaultIdentifier => "edit_default_identifier",
+        UiAction::ShowOptionsOverview => "show_options_overview",
+        UiAction::PickTrashFolder => "pick_trash_folder",
+        UiAction::OpenTrashFolder => "open_trash_folder",
+        UiAction::OpenPrimaryGroupPicker => "open_primary_group_picker",
+        UiAction::OpenPrimaryItemPicker => "open_primary_item_picker",
+        UiAction::OpenPrimaryNumberPicker => "open_primary_number_picker",
+        UiAction::OpenSecondaryGroupPicker => "open_secondary_group_picker",
+        UiAction::OpenSecondaryItemPicker => "open_secondary_item_picker",
+        UiAction::OpenSecondaryNumberPicker => "open_secondary_number_picker",
+        UiAction::SetPrimaryGroup { .. } => "set_primary_group",
+        UiAction::SetPrimaryItem { .. } => "set_primary_item",
+        UiAction::SetPrimaryNumber { .. } => "set_primary_number",
+        UiAction::SetSecondaryGroup { .. } => "set_secondary_group",
+        UiAction::SetSecondaryItem { .. } => "set_secondary_item",
+        UiAction::SetSecondaryNumber { .. } => "set_secondary_number",
+        UiAction::FocusFolderSearch => "focus_folder_search",
+        UiAction::SetFolderSearch { .. } => "set_folder_search",
+        UiAction::ToggleShowAllFolders => "toggle_show_all_folders",
+        UiAction::ToggleFolderFlattenedView => "toggle_folder_flattened_view",
+        UiAction::SelectSourceRow { .. } => "select_source_row",
+        UiAction::FocusSourceRow { .. } => "focus_source_row",
+        UiAction::MoveSourceFocus { .. } => "move_source_focus",
+        UiAction::ReloadFocusedSourceRow => "reload_focused_source_row",
+        UiAction::HardSyncFocusedSourceRow => "hard_sync_focused_source_row",
+        UiAction::OpenFocusedSourceFolder => "open_focused_source_folder",
+        UiAction::RemoveFocusedSourceRow => "remove_focused_source_row",
+        UiAction::ReloadSourceRow { .. } => "reload_source_row",
+        UiAction::HardSyncSourceRow { .. } => "hard_sync_source_row",
+        UiAction::OpenSourceFolderRow { .. } => "open_source_folder_row",
+        UiAction::RemoveSourceRow { .. } => "remove_source_row",
+        UiAction::FocusFolderRow { .. } => "focus_folder_row",
+        UiAction::ActivateFolderRow { .. } => "activate_folder_row",
+        UiAction::ToggleFolderRowExpanded { .. } => "toggle_folder_row_expanded",
+        UiAction::ExpandFocusedFolder => "expand_focused_folder",
+        UiAction::CollapseFocusedFolder => "collapse_focused_folder",
+        UiAction::ToggleFocusedFolderSelection => "toggle_focused_folder_selection",
+        UiAction::MoveFolderFocus { .. } => "move_folder_focus",
+        UiAction::StartNewFolder => "start_new_folder",
+        UiAction::StartNewFolderAtFolderRow { .. } => "start_new_folder_at_folder_row",
+        UiAction::StartNewFolderAtRoot => "start_new_folder_at_root",
+        UiAction::FocusFolderCreateInput => "focus_folder_create_input",
+        UiAction::SetFolderCreateInput { .. } => "set_folder_create_input",
+        UiAction::ConfirmFolderCreate => "confirm_folder_create",
+        UiAction::CancelFolderCreate => "cancel_folder_create",
+        UiAction::StartFolderRename => "start_folder_rename",
+        UiAction::DeleteFocusedFolder => "delete_focused_folder",
+        UiAction::RestoreRetainedFolderDeletes => "restore_retained_folder_deletes",
+        UiAction::PurgeRetainedFolderDeletes => "purge_retained_folder_deletes",
+        UiAction::ClearFolderDeleteRecoveryLog => "clear_folder_delete_recovery_log",
+        _ => return None,
+    })
+}
