+++ src/app_core/native_shell/composition/state/automation/browser.rs
@@ -1,8 +1,7 @@
 //! Browser-panel automation snapshot builders.
 
-use super::helpers::{
-    action_slug, bool_text, bounds, circle_rect, metadata, node_id, simple_node, slug,
-};
+use super::action_slugs::action_slug;
+use super::helpers::{bool_text, bounds, circle_rect, metadata, node_id, simple_node, slug};
 use super::*;
 use crate::app_core::native_shell::runtime_contract::AutomationRole;
