+++ src/app_core/native_shell/composition/state/automation/waveform.rs
@@ -1,6 +1,7 @@
 //! Waveform automation snapshot builders.
 
-use super::helpers::{action_slug, bounds, metadata, node_id, selection_micros_text, simple_node};
+use super::action_slugs::action_slug;
+use super::helpers::{bounds, metadata, node_id, selection_micros_text, simple_node};
 use super::*;
 use crate::app_core::native_shell::runtime_contract::{AutomationRole, NormalizedRangeModel};
